Add dashboard UI for progress portal and some css
The logo is in a block so that we can switch between logo versions in
different templates. Fixed css for navbar title. Fixed app display area,
which had a align_items = center which was causing an error for other
divs under it, by removing it. New svg for progress_portal dashboard.