Add responsive secondary navbar system for progress portal
Copy navbar_toggle.css and navbar_toggle.js to the main dashboard of a
new app.
Add .secondary_navbar class to the container div for the navbar
And set other responsive parameters
We now have a easy way to create toggle feature for navbars

Todo: Add similar funcitonality for creation of secondary navbar in
general
Also find a way to contain elements properly inside the divs in smaller
screen. For example: When the text is too long, the display shrinks and
reveals background.